STATUS-4 - Revamp
Doing a final search for currentTimeMillis, turned up an additional usage on the core SystemReport.

Threading the WallClock references through the system reporter and down to the check result set ended
up being a fair number of changes but slightly cleaner in the end, since the SystemReporter itself
can be the sole owner of the wallclock instance.