STATUS-26: Set interrupted flag in DependencyChecker
Set interrupted flag on InterruptedException in DependencyChecker.
It is the correct thing to do and dong so does not seem to break any
of the existing logic.

The comment saying not to do that appears to be written by somebody
assuming InterruptedException is thrown on a per-future basis, which
is definitely not the case. InterruptedException is thrown from blocking
methods such as Future.get() if the *calling* thread is interrupted.